+++ src/lib/server/db/schema/index.ts
@@ -17,6 +17,7 @@ export * from "./shareLink";
 export * from "./song";
 export * from "./songDocVersion";
 export * from "./stem";
+export * from "./twoFactor";
 export * from "./user";
 export * from "./userDoc";
 export * from "./userDocVersion";